added inert_stochastic_gradient_descent

+++ src/linear_algebra.rs
@@ -5,9 +5,9 @@ use num::One;
 pub trait Vector : Clone{
 
     type Scalar : Num + Zero + One + Copy;
-    fn dimension(&self) -> u32;
-    fn at(&self, i : u32) -> Self::Scalar;
-    fn mut_at<'a>(&'a mut self, i : u32) -> &'a mut Self::Scalar;
+    fn dimension(&self) -> usize;
+    fn at(&self, i : usize) -> Self::Scalar;
+    fn mut_at<'a>(&'a mut self, i : usize) -> &'a mut Self::Scalar;
     // fn dot(&self, other : &Self) -> Self::Scalar;
     fn dot(&self, other : &Self) -> Self::Scalar{
         let mut result = Self::Scalar::zero();
@@ -21,15 +21,15 @@ pub trait Vector : Clone{
 impl Vector for f64{
     type Scalar = f64;
 
-    fn dimension(&self) -> u32{
+    fn dimension(&self) -> usize{
         1
     }
 
-    fn at(&self, _ : u32) -> f64{
+    fn at(&self, _ : usize) -> f64{
         *self
     }
 
-    fn mut_at<'a>(&'a mut self, _ : u32) -> &'a mut f64{
+    fn mut_at<'a>(&'a mut self, _ : usize) -> &'a mut f64{
         self
     }
 
@@ -41,16 +41,16 @@ impl Vector for f64{
 impl Vector for [f64;2]{
     type Scalar = f64;
 
-    fn dimension(&self) -> u32{
+    fn dimension(&self) -> usize{
         2
     }
 
-    fn at(&self, index : u32) -> f64{
-        self[index as usize]
+    fn at(&self, index : usize) -> f64{
+        self[index]
     }
 
-    fn mut_at<'a>(&'a mut self, index : u32) -> &'a mut f64{
-        & mut self[index as usize]
+    fn mut_at<'a>(&'a mut self, index : usize) -> &'a mut f64{
+        & mut self[index]
     }
 }
